WCMS-25325: Move the DatasetListSubmenu component into the components folder.
